requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
python-multipart==0.0.20
aiofiles==24.1.0
websockets==15.0.1
# sentence-transformers==3.3.1  # REMOVED: Heavy embedding models (300MB+)
scikit-learn==1.6.0
pypdf2==3.0.1
python-dotenv==1.0.1
httpx==0.28.1
aiohttp==3.12.13
nest-asyncio==1.6.0
jupyter==1.1.1
ipykernel==6.29.5

# Database
sqlalchemy==2.0.36
# Note: sqlite3 is built into Python, no need to install

# PDF Processing
pymupdf==1.24.13

# Research APIs
arxiv==2.1.3
biopython==1.84

# Enhanced LLM Support (optional - only if using external APIs)
# openai==1.57.2
# anthropic==0.40.0

# Text Processing (lightweight alternatives)
# transformers==4.47.1  # REMOVED: Heavy transformer models (1GB+)
# torch==2.5.1          # REMOVED: PyTorch dependency (500MB+)
nltk==3.9.1             # Lightweight NLP library
certifi==2024.8.30

# Free Text-to-Speech Alternatives (lightweight)
gtts==2.5.4          # Google Text-to-Speech (free, requires internet)
pyttsx3==2.90        # Offline TTS (free, no internet required)
